Move `account_info_storage_key` to `polkadot-core` crate
