+++ primitives/chain-rococo/Cargo.toml
@@ -15,14 +15,10 @@ bp-polkadot-core = { path = "../polkadot-core", default-features = false }
 bp-runtime = { path = "../runtime", default-features = false }
 
 # Substrate Based Dependencies
-frame-support = { git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
 sp-api = { git = "https://github.com/paritytech/substrate", branch = "master" , default-features = false }
 sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
 sp-std = { git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
 sp-version = { git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
-
-[dev-dependencies]
-hex = "0.4"
 
 [features]
 default = ["std"]
@@ -30,7 +26,6 @@ std = [
 	"bp-messages/std",
 	"bp-polkadot-core/std",
 	"bp-runtime/std",
-	"frame-support/std",
 	"parity-scale-codec/std",
 	"sp-api/std",
 	"sp-runtime/std",
