+++ primitives/polkadot-core/src/lib.rs
@@ -25,7 +25,7 @@ use frame_support::{
 		constants::{BlockExecutionWeight, ExtrinsicBaseWeight, WEIGHT_PER_SECOND},
 		DispatchClass, Weight,
 	},
-	RuntimeDebug,
+	Blake2_128Concat, RuntimeDebug, StorageHasher, Twox128,
 };
 use frame_system::limits;
 use sp_core::Hasher as HasherT;
@@ -34,6 +34,7 @@ use sp_runtime::{
 	traits::{BlakeTwo256, IdentifyAccount, Verify},
 	MultiSignature, OpaqueExtrinsic, Perbill,
 };
+use sp_std::prelude::Vec;
 
 // Re-export's to avoid extra substrate dependencies in chain-specific crates.
 pub use frame_support::Parameter;
@@ -290,6 +291,26 @@ impl Convert<sp_core::H256, AccountId> for AccountIdConverter {
 	}
 }
 
+/// Return a storage key for account data.
+///
+/// This is based on FRAME storage-generation code from Substrate:
+/// https://github.com/paritytech/substrate/blob/c939ceba381b6313462d47334f775e128ea4e95d/frame/support/src/storage/generator/map.rs#L74
+/// The equivalent command to invoke in case full `Runtime` is known is this:
+/// `let key = frame_system::Account::<Runtime>::storage_map_final_key(&account_id);`
+pub fn account_info_storage_key(id: &AccountId) -> Vec<u8> {
+	let module_prefix_hashed = Twox128::hash(b"System");
+	let storage_prefix_hashed = Twox128::hash(b"Account");
+	let key_hashed = parity_scale_codec::Encode::using_encoded(id, Blake2_128Concat::hash);
+
+	let mut final_key = Vec::with_capacity(module_prefix_hashed.len() + storage_prefix_hashed.len() + key_hashed.len());
+
+	final_key.extend_from_slice(&module_prefix_hashed[..]);
+	final_key.extend_from_slice(&storage_prefix_hashed[..]);
+	final_key.extend_from_slice(&key_hashed);
+
+	final_key
+}
+
 #[cfg(test)]
 mod tests {
 	use super::*;
@@ -304,5 +325,16 @@ mod tests {
 			actual_size,
 			MAXIMAL_ENCODED_ACCOUNT_ID_SIZE,
 		);
+	}
+
+	#[test]
+	fn should_generate_storage_key() {
+		let acc = [
+			1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
+			30, 31, 32,
+		]
+		.into();
+		let key = account_info_storage_key(&acc);
+		assert_eq!(hex::encode(key), "26aa394eea5630e07c48ae0c9558cef7b99d880ec681799c0cf30e8886371da92dccd599abfe1920a1cff8a7358231430102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20");
 	}
 }
